Release candidate of v1.5, compare to v1.3: - Codebase fully refactored, rewritten using modern C++, much more maintainable. - Added option to use Win+Backspace as hotkey. - Click on Notify Icon will show status in 'balloon' notification. - On unexpected error, it will show error message instead of rage quit/ignoring.
Downsides:
* Codebase itself still support Windows 2000 but it's compiled using MSVC2015 without XP targeting stuff so...
* Codebase has 0% compatibility with VC6, so no more 'ver tiny', you will need VC2015 redist to run it.

I will see if I can compile it using TDM-GCC later, to get it runs under Win2K/XP.

+++ noborder/NotifyIcon.cpp
@@ -38,7 +38,9 @@ LRESULT NotifyIcon::WndProc(HWND hwnd, UINT msg, WPARAM w, LPARAM l) {
 		auto self = reinterpret_cast<NotifyIcon*>(ptr);
 		if (msg == NI_CALLBACK_MSG) {
 			/* wparam = id, lparam = wm_xxx */
-			if (self->m_eventHandler) { self->m_eventHandler(*self, l); }
+			if (self->m_eventHandler) {
+				self->m_eventHandler(*self, static_cast<UINT>(l));
+			}
 		}
 		else if (msg == self->m_taskbarCreatedMsg) {
 			self->Update(true);
